+++ user/pfq-lang/src/Paths.hs
@@ -15,7 +15,7 @@
 -- Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
 --
 
-module Paths_qlang where
+module Paths_pfq_lang where
 
 import Data.Version(Version(..))